misc: rename NewCommonPlayerViewController.swift
